Struct aws_sdk_secretsmanager::client::fluent_builders::CreateSecret

pub struct CreateSecret<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateSecret<C>

pub async fn send(
self
) -> Result<CreateSecretOutput, SdkError<CreateSecretError>> where
C: SmithyConnector,

pub fn name(self, input: impl Into<String>) -> Self

Specifies the friendly name of the new secret.

The secret name must be ASCII letters, digits, or the following characters : /_+=.@-

Do not end your secret name with a hyphen followed by six characters. If you do so, you risk confusion and unexpected results when searching for a secret by partial ARN. Secrets Manager automatically adds a hyphen and six random characters at the end of the ARN.

pub fn set_name(self, input: Option<String>) -> Self

pub fn client_request_token(self, input: impl Into<String>) -> Self

(Optional) If you include SecretString or SecretBinary, then an initial version is created as part of the secret, and this parameter specifies a unique identifier for the new version.

If you use the AWS CLI or one of the AWS SDK to call this operation, then you can leave this parameter empty. The CLI or SDK generates a random UUID for you and includes it as the value for this parameter in the request. If you don't use the SDK and instead generate a raw HTTP request to the Secrets Manager service endpoint, then you must generate a ClientRequestToken yourself for the new version and include the value in the request.

This value helps ensure idempotency. Secrets Manager uses this value to prevent the accidental creation of duplicate versions if there are failures and retries during a rotation. We recommend that you generate a UUID-type value to ensure uniqueness of your versions within the specified secret.

If the ClientRequestToken value isn't already associated with a version of the secret then a new version of the secret is created.

If a version with this value already exists and the version SecretString and SecretBinary values are the same as those in the request, then the request is ignored.

If a version with this value already exists and that version's SecretString and SecretBinary values are different from those in the request, then the request fails because you cannot modify an existing version. Instead, use PutSecretValue to create a new version.

This value becomes the VersionId of the new version.

pub fn set_client_request_token(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

(Optional) Specifies a user-provided description of the secret.

pub fn set_description(self, input: Option<String>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

(Optional) Specifies the ARN, Key ID, or alias of the AWS KMS customer master key (CMK) to be used to encrypt the SecretString or SecretBinary values in the versions stored in this secret.

You can specify any of the supported ways to identify a AWS KMS key ID. If you need to reference a CMK in a different account, you can use only the key ARN or the alias ARN.

If you don't specify this value, then Secrets Manager defaults to using the AWS account's default CMK (the one named aws/secretsmanager). If a AWS KMS CMK with that name doesn't yet exist, then Secrets Manager creates it for you automatically the first time it needs to encrypt a version's SecretString or SecretBinary fields.

You can use the account default CMK to encrypt and decrypt only if you call this operation using credentials from the same account that owns the secret. If the secret resides in a different account, then you must create a custom CMK and specify the ARN in this field.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn secret_binary(self, input: Blob) -> Self

(Optional) Specifies binary data that you want to encrypt and store in the new version of the secret. To use this parameter in the command-line tools, we recommend that you store your binary data in a file and then use the appropriate technique for your tool to pass the contents of the file as a parameter.

Either SecretString or SecretBinary must have a value, but not both. They cannot both be empty.

This parameter is not available using the Secrets Manager console. It can be accessed only by using the AWS CLI or one of the AWS SDKs.

pub fn set_secret_binary(self, input: Option<Blob>) -> Self

pub fn secret_string(self, input: impl Into<String>) -> Self

(Optional) Specifies text data that you want to encrypt and store in this new version of the secret.

Either SecretString or SecretBinary must have a value, but not both. They cannot both be empty.

If you create a secret by using the Secrets Manager console then Secrets Manager puts the protected secret text in only the SecretString parameter. The Secrets Manager console stores the information as a JSON structure of key/value pairs that the Lambda rotation function knows how to parse.

For storing multiple values, we recommend that you use a JSON text string argument and specify key/value pairs. For information on how to format a JSON parameter for the various command line tool environments, see Using JSON for Parameters in the AWS CLI User Guide. For example:

{"username":"bob","password":"abc123xyz456"}

If your command-line tool or SDK requires quotation marks around the parameter, you should use single quotes to avoid confusion with the double quotes required in the JSON text.

pub fn set_secret_string(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

(Optional) Specifies a list of user-defined tags that are attached to the secret. Each tag is a "Key" and "Value" pair of strings. This operation only appends tags to the existing list of tags. To remove tags, you must use UntagResource.

Secrets Manager tag key names are case sensitive. A tag with the key "ABC" is a different tag from one with key "abc".

If you check tags in IAM policy Condition elements as part of your security strategy, then adding or removing a tag can change permissions. If the successful completion of this operation would result in you losing your permissions for this secret, then this operation is blocked and returns an Access Denied error.

This parameter requires a JSON text string argument. For information on how to format a JSON parameter for the various command line tool environments, see Using JSON for Parameters in the AWS CLI User Guide. For example:

[{"Key":"CostCenter","Value":"12345"},{"Key":"environment","Value":"production"}]

If your command-line tool or SDK requires quotation marks around the parameter, you should use single quotes to avoid confusion with the double quotes required in the JSON text.

The following basic restrictions apply to tags:

Maximum number of tags per secret—50

Maximum key length—127 Unicode characters in UTF-8

Maximum value length—255 Unicode characters in UTF-8

Tag keys and values are case sensitive.

Do not use the aws: prefix in your tag names or values because AWS reserves it for AWS use. You can't edit or delete tag names or values with this prefix. Tags with this prefix do not count against your tags per secret limit.

If you use your tagging schema across multiple services and resources, remember other services might have restrictions on allowed characters. Generally allowed characters: letters, spaces, and numbers representable in UTF-8, plus the following special characters: + - = . _ : / @.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn add_replica_regions(self, inp: impl Into<ReplicaRegionType>) -> Self

(Optional) Add a list of regions to replicate secrets. Secrets Manager replicates the KMSKeyID objects to the list of regions specified in the parameter.

pub fn set_add_replica_regions(
self,
input: Option<Vec<ReplicaRegionType>>
) -> Self

pub fn force_overwrite_replica_secret(self, input: bool) -> Self

(Optional) If set, the replication overwrites a secret with the same name in the destination region.

pub fn set_force_overwrite_replica_secret(self, input: Option<bool>) -> Self