Module aws_sdk_secretsmanager::error

Modules

decryption_failure

See DecryptionFailure

encryption_failure

See EncryptionFailure

internal_service_error

See InternalServiceError

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_exception

See InvalidParameterException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

malformed_policy_document_exception

See MalformedPolicyDocumentException

precondition_not_met_exception

See PreconditionNotMetException

public_policy_exception

See PublicPolicyException

resource_exists_exception

See ResourceExistsException

resource_not_found_exception

See ResourceNotFoundException

Structs

CancelRotateSecretError

CreateSecretError

DecryptionFailure

Secrets Manager can't decrypt the protected secret text using the provided KMS key.

DeleteResourcePolicyError

DeleteSecretError

DescribeSecretError

EncryptionFailure

Secrets Manager can't encrypt the protected secret text using the provided KMS key. Check that the

GetRandomPasswordError

GetResourcePolicyError

GetSecretValueError

InternalServiceError

An error occurred on the server side.

InvalidNextTokenException

You provided an invalid NextToken value.

InvalidParameterException

You provided an invalid value for a parameter.

InvalidRequestException

You provided a parameter value that is not valid for the current state of the

LimitExceededException

The request failed because it would exceed one of the Secrets Manager internal limits.

ListSecretVersionIdsError

ListSecretsError

MalformedPolicyDocumentException

You provided a resource-based policy with syntax errors.

PreconditionNotMetException

The request failed because you did not complete all the prerequisite steps.

PublicPolicyException

The BlockPublicPolicy parameter is set to true and the resource policy did not prevent broad access to the secret.

PutResourcePolicyError

PutSecretValueError

RemoveRegionsFromReplicationError

ReplicateSecretToRegionsError

ResourceExistsException

A resource with the ID you requested already exists.

ResourceNotFoundException

We can't find the resource that you asked for.

RestoreSecretError

RotateSecretError

StopReplicationToReplicaError

TagResourceError

UntagResourceError

UpdateSecretError

UpdateSecretVersionStageError

ValidateResourcePolicyError

Enums

CancelRotateSecretErrorKind

CreateSecretErrorKind

DeleteResourcePolicyErrorKind

DeleteSecretErrorKind

DescribeSecretErrorKind

GetRandomPasswordErrorKind

GetResourcePolicyErrorKind

GetSecretValueErrorKind

ListSecretVersionIdsErrorKind

ListSecretsErrorKind

PutResourcePolicyErrorKind

PutSecretValueErrorKind

RemoveRegionsFromReplicationErrorKind

ReplicateSecretToRegionsErrorKind

RestoreSecretErrorKind

RotateSecretErrorKind

StopReplicationToReplicaErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateSecretErrorKind

UpdateSecretVersionStageErrorKind

ValidateResourcePolicyErrorKind