Module aws_sdk_secretsmanager::model

Modules

filter

See Filter

replica_region_type

See ReplicaRegionType

replication_status_type

See ReplicationStatusType

rotation_rules_type

See RotationRulesType

secret_list_entry

See SecretListEntry

secret_versions_list_entry

See SecretVersionsListEntry

tag

See Tag

validation_errors_entry

See ValidationErrorsEntry

Structs

Filter

Allows you to add filters when you use the search function in Secrets Manager.

ReplicaRegionType

(Optional) Custom type consisting of a Region (required) and the KmsKeyId which can be an ARN, Key ID, or Alias.

ReplicationStatusType

A replication object consisting of a RegionReplicationStatus object and includes a Region, KMSKeyId, status, and status message.

RotationRulesType

A structure that defines the rotation configuration for the secret.

SecretListEntry

A structure that contains the details about a secret. It does not include the encrypted

SecretVersionsListEntry

A structure that contains information about one version of a secret.

Tag

A structure that contains information about a tag.

ValidationErrorsEntry

Displays errors that occurred during validation of the resource policy.

Enums

FilterNameStringType

SortOrderType

StatusType