Module aws_sdk_secretsmanager::operation

Structs

CancelRotateSecret
Disables automatic scheduled rotation and cancels the rotation of a secret if currently in

CreateSecret
Creates a new secret. A secret in Secrets Manager consists of both the protected secret data and the

DeleteResourcePolicy
Deletes the resource-based permission policy attached to the secret.

DeleteSecret
Deletes an entire secret and all of its versions. You can optionally include a recovery

DescribeSecret
Retrieves the details of a secret. It does not include the encrypted fields. Secrets

GetRandomPassword
Generates a random password of the specified complexity. This operation is intended for

GetResourcePolicy
Retrieves the JSON text of the resource-based policy document attached to the

GetSecretValue
Retrieves the contents of the encrypted fields SecretString or

ListSecretVersionIds
Lists all of the versions attached to the specified secret. The output does not include

ListSecrets
Lists all of the secrets that are stored by Secrets Manager in the AWS account. To list the

PutResourcePolicy
Attaches the contents of the specified resource-based permission policy to a secret. A

PutSecretValue
Stores a new encrypted secret value in the specified secret. To do this, the operation

RestoreSecret
Cancels the scheduled deletion of a secret by removing the DeletedDate time

RotateSecret
Configures and starts the asynchronous process of rotating this secret. If you include the

TagResource
Attaches one or more tags, each consisting of a key name and a value, to the specified

UntagResource
Removes one or more tags from the specified secret.

UpdateSecret
Modifies many of the details of the specified secret. If you include a

UpdateSecretVersionStage
Modifies the staging labels attached to a version of a secret. Staging labels are used to

ValidateResourcePolicy
Validates the JSON text of the resource-based policy document attached to the