Module aws_sdk_secretsmanager::error

Modules

decryption_failure
See DecryptionFailure

encryption_failure
See EncryptionFailure

internal_service_error
See InternalServiceError

invalid_next_token_error
See InvalidNextTokenError

invalid_parameter_error
See InvalidParameterError

invalid_request_error
See InvalidRequestError

limit_exceeded_error
See LimitExceededError

malformed_policy_document_error
See MalformedPolicyDocumentError

precondition_not_met_error
See PreconditionNotMetError

public_policy_error
See PublicPolicyError

resource_exists_error
See ResourceExistsError

resource_not_found_error
See ResourceNotFoundError

Structs

CancelRotateSecretError

CreateSecretError

DecryptionFailure
Secrets Manager can't decrypt the protected secret text using the provided KMS key.

DeleteResourcePolicyError

DeleteSecretError

DescribeSecretError

EncryptionFailure
Secrets Manager can't encrypt the protected secret text using the provided KMS key. Check that the

GetRandomPasswordError

GetResourcePolicyError

GetSecretValueError

InternalServiceError
An error occurred on the server side.

InvalidNextTokenError
You provided an invalid NextToken value.

InvalidParameterError
You provided an invalid value for a parameter.

InvalidRequestError
You provided a parameter value that is not valid for the current state of the

LimitExceededError
The request failed because it would exceed one of the Secrets Manager internal limits.

ListSecretVersionIdsError

ListSecretsError

MalformedPolicyDocumentError
The policy document that you provided isn't valid.

PreconditionNotMetError
The request failed because you did not complete all the prerequisite steps.

PublicPolicyError
The resource policy did not prevent broad access to the secret.

PutResourcePolicyError

PutSecretValueError

ResourceExistsError
A resource with the ID you requested already exists.

ResourceNotFoundError
We can't find the resource that you asked for.

RestoreSecretError

RotateSecretError

TagResourceError

UntagResourceError

UpdateSecretError

UpdateSecretVersionStageError

ValidateResourcePolicyError

Enums

CancelRotateSecretErrorKind

CreateSecretErrorKind

DeleteResourcePolicyErrorKind

DeleteSecretErrorKind

DescribeSecretErrorKind

GetRandomPasswordErrorKind

GetResourcePolicyErrorKind

GetSecretValueErrorKind

ListSecretVersionIdsErrorKind

ListSecretsErrorKind

PutResourcePolicyErrorKind

PutSecretValueErrorKind

RestoreSecretErrorKind

RotateSecretErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateSecretErrorKind

UpdateSecretVersionStageErrorKind

ValidateResourcePolicyErrorKind